refactor(frontend): move isSharedAccount to src/utils/sharedAccount
